combined related queries to domains, simplified function names
