+++ Webapp/sources/services/queries/__init__.py
@@ -1,9 +1,1 @@
-from . import (
-    compound,
-    compound_data_error_report,
-    reaction,
-    user,
-    workbook,
-    workgroup,
-    workgroup_request,
-)
+from . import compound, reaction, user, workbook, workgroup
